requirements.txt
email-validator==2.2.0
flask-restful==0.3.10
flask==3.1.0
flask-sqlalchemy==3.1.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
flask-login==0.6.3
sqlalchemy==2.0.40
werkzeug==3.1.3
geopy==2.4.1
marshmallow==3.26.1
Flask-Cors==4.0.1
Flask-Limiter==3.6.0
python-dotenv==1.0.1
supabase==2.5.1
Flask-Migrate==4.1.0
flask-jwt-extended==4.6.0
fastapi
uvicorn[standard]
google-auth==2.29.0
google-auth-oauthlib==1.1.0
requests==2.31.0